source/auth/requirements.txt
fastapi[standard]
uvicorn[standard]
bcrypt
pyjwt
email-validator
python-dotenv
psycopg2
sqlalchemy
